Adicionando o component de mascarano sistema e mudando a variavel preco para double(back-end)

+++ ftt/src/main/java/br/com/unievangelica/ftt/domain/quadra/Quadra.java
@@ -24,7 +24,7 @@ public class Quadra extends AbstractEntity {
     private String nome;
 
     @Column(name = "preco", nullable = false)
-    private int preco;
+    private double preco;
 
     @NotNull
     @ManyToOne
